Avoid i386 compile error related to comparison
Adds casts to uint64_t so that a comparison between size_t and the
largest possible uint32_t value does not lead to a tautological warning
when building for i386.

Fixes tint:1162

Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/63440
Auto-Submit: Alastair Donaldson <afdx@google.com>
Commit-Queue: Alastair Donaldson <afdx@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>

+++ src/inspector/inspector.cc
@@ -321,7 +321,8 @@ uint32_t Inspector::GetStorageSize(const std::string& entry_point) {
     }
   }
 
-  if (size > std::numeric_limits<uint32_t>::max()) {
+  if (static_cast<uint64_t>(size) >
+      static_cast<uint64_t>(std::numeric_limits<uint32_t>::max())) {
     return std::numeric_limits<uint32_t>::max();
   }
   return static_cast<uint32_t>(size);
